requirements.txt
Django==5.1.3
pillow==11.0.0
dj-svg==0.3.1
djangorestframework=3.15.2